DPL: separate arrow related services in own file (#5789)
Propedeutic for some bigger optimization to the rate limiter.
