+++ Framework/Core/src/ArrowSupport.h
@@ -7,8 +7,8 @@
 // In applying this license CERN does not waive the privileges and immunities
 // granted to it by virtue of its status as an Intergovernmental Organization
 // or submit itself to any jurisdiction.
-#ifndef O2_FRAMEWORK_COMMONMESSAGEBACKENDS_H_
-#define O2_FRAMEWORK_COMMONMESSAGEBACKENDS_H_
+#ifndef O2_FRAMEWORK_ARROWSUPPORT_H_
+#define O2_FRAMEWORK_ARROWSUPPORT_H_
 
 #include "Framework/ServiceSpec.h"
 #include "Framework/TypeIdHelpers.h"
@@ -17,12 +17,13 @@ namespace o2::framework
 {
 
 /// A few ServiceSpecs data sending backends
-struct CommonMessageBackends {
-  static ServiceSpec fairMQBackendSpec();
-  static ServiceSpec stringBackendSpec();
-  static ServiceSpec rawBufferBackendSpec();
+struct ArrowSupport {
+  // Rate limiting service
+  static ServiceSpec rateLimitingSpec();
+  // Create spec for backend used to send Arrow messages
+  static ServiceSpec arrowBackendSpec();
 };
 
 } // namespace o2::framework
 
-#endif // O2_FRAMEWORK_COMMONMESSAGEBACKENDS_H_
+#endif // O2_FRAMEWORK_ARROWSUPPORT_H_
